Skapat fil docker-compose.yml och även klass "Ownership". Även fixat …

+++ Padel/src/main/java/com/strom/wigellPadel/repositories/CourtRepo.java
@@ -3,5 +3,5 @@ package com.strom.wigellPadel.repositories;
 import com.strom.wigellPadel.entities.Court;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CourtRepo extends JpaRepository<Court, Integer> {
+public interface CourtRepo extends JpaRepository<Court, Long> {
 }
